Add files via upload
In this project the dataset of zomato is first preprocessed and then used for the analysis. The insights are visualized and the outcome provides the deep information about the restaurant it's rating , cuisines and location too. It provides the data insights of online order and the budget required for 2 people . It helps to choose restaurant according to the rating location and budget.